requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
bcrypt==4.2.0
beanie==1.27.0
certifi==2024.8.30
click==8.1.7
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.6.0
openai==1.51.2
orjson==3.10.7
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.1.1
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
toml==0.10.2
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
